CRUD de administrador para medico completo y totalmente funcional

+++ src/main/java/evovital/uniquindio/edu/co/dto/medico/MedicoDTOAdmin.java
@@ -2,7 +2,7 @@ package evovital.uniquindio.edu.co.dto.medico;
 
 public record MedicoDTOAdmin(
 
-        int codigo,
+        Long codigo,
         String nombre,
         String urlFoto,
         String especialidad
